DataPoint refactoring done
