+++ openTSDB.net.Test/DataPointFixtures/DataPointListSample.json
@@ -1,1 +1,1 @@
-[{"metric":"testMetric","timestamp":1449878400,"value":15,"tags":{"host":"testHost"}},{"metric":"testMetric","timestamp":1449878400,"value":14,"tags":{"host":"testHost"}}]
+[{"metric":"testMetric","timestamp":1449878400,"value":15,"tags":{"host":"testHost","host_artificial":"true"}},{"metric":"testMetric","timestamp":1449878400,"value":14,"tags":{"host":"testHost","host_artificial":"true"}}]
