+++ openTSDB.net/OpenTsdbManager.cs
@@ -24,7 +24,7 @@ namespace OpenTsdbNet
             {
                 throw new ArgumentNullException(ErrorMessages.DATA_POINT_NULL, nameof(dataPoint));
             }
-            return await TsdbServer.PublishDataAsync(dataPoint.Bytify());
+            return await TsdbServer.PublishDataAsync(dataPoint.AsByteArray());
 
         }
 
@@ -34,7 +34,7 @@ namespace OpenTsdbNet
             {
                 throw new ArgumentNullException(ErrorMessages.DATA_POINT_NULL, nameof(dataPoints));
             }
-            return await TsdbServer.PublishDataAsync(dataPoints.Bytify());
+            return await TsdbServer.PublishDataAsync(dataPoints.AsByteArray());
         }
 
         public async Task<TsdbSubmissionResponse> PushAsync<T>(string name, T value)
